requirements.txt
discord.py>=2.0.0
python-dotenv>=0.19.0
openai>=1.0.0
pytest>=7.0.0
pytest-cov>=4.0.0
